+++ prost-reflect-tests/build.rs
@@ -68,7 +68,7 @@ fn main() -> io::Result<()> {
                 "src/desc.proto",
                 "src/desc2.proto",
                 "src/desc_no_package.proto",
-                "src/import_any.proto",
+                "src/imports.proto",
             ],
             &["src/"],
         )?;